Update patch set 1
Patch Set 1:

> is this a duplicate of this one here? https://review.rdoproject.org/r/#/c/10681/

No, this is not a duplicate. Or better, this one is an alignment to the build shipped in RHV / RHOSE. The other one is the tentative to make it passing fedora guidelines (but it's currently failing)

 > Is your -2 from there still true here?

No, it's not valid here. This seems to work fine (but not passing fedora guidelines)

Patch-set: 1
Label: Code-Review=0